ian/creating_backend_abstraction_for_self_hosted_worker
